Do not acquire vbus cap in library.
Move the acquisition of the vbus capability necessary for enumerating the
device bus handed to the driver out of the library. The driver library shall
not make any hardcoded assumptions on the name of capabilities that it needs
for working correctly, since those may be different for each application that
this package is used in.