API: Group initial dev config in dedicated struct.
Changes the API for setting up an Ixl device by moving the set of available
device configuration parameters to a dedicated structure. This reduces the
number of arguments to the ixl_init function and provides applications with
a reasonable defaul configuration. This change is a preparatory step for
enabling Ixl to support more sophisticated hardware features.