Fix IRQ unmasking in interrupt wait mode.
If the driver is configured to run in interrupt wait mode, it should take
care full responsibility of IRQ handling, including automatic masking and
unmasking of interrupts. This is different from the notify IRQ mode where
applications perform these steps manually.

The previous IRQ rework missed unmasking of receive interrupts if the driver
was configured to run in interrupt wait mode. This caused severe performance
degradation of applications that use the IRQ wait mode.